+++ tests/Command/DaemonCommandTest.php
@@ -1,12 +1,17 @@
 <?php
 
-namespace Phlib\ConsoleProcess\Tests\Command;
+declare(strict_types=1);
 
-use Symfony\Component\Console\Tester\CommandTester;
-use Symfony\Component\Console\Application;
+namespace Phlib\ConsoleProcess\Command;
+
 use phpmock\phpunit\PHPMock;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Tester\CommandTester;
 
-class DaemonCommandTest extends \PHPUnit_Framework_TestCase
+class DaemonCommandTest extends TestCase
 {
     use PHPMock;
 
@@ -21,7 +26,7 @@ class DaemonCommandTest extends \PHPUnit_Framework_TestCase
     protected $commandName = 'foo:bar';
 
     /**
-     * @var \TestCommand|\PHPUnit_Framework_MockObject_MockObject
+     * @var Stub\DaemonCommandStub|MockObject
      */
     protected $command;
 
@@ -30,17 +35,12 @@ class DaemonCommandTest extends \PHPUnit_Framework_TestCase
      */
     protected $tester;
 
-    public static function setUpBeforeClass()
-    {
-        require_once __DIR__ . '/files/DaemonCommandStub.php';
-    }
-
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
         $this->application = new Application();
-        $this->application->add(new \DaemonCommandStub($this->commandName));
+        $this->application->add(new Stub\DaemonCommandStub($this->commandName));
 
         $this->command = $this->application->find($this->commandName);
         $this->tester = new CommandTester($this->command);
@@ -49,7 +49,7 @@ class DaemonCommandTest extends \PHPUnit_Framework_TestCase
         });
     }
 
-    public function tearDown()
+    protected function tearDown(): void
     {
         $this->tester = null;
         $this->command = null;
@@ -57,166 +57,171 @@ class DaemonCommandTest extends \PHPUnit_Framework_TestCase
         parent::tearDown();
     }
 
-    public function testInstanceOfConsoleCommand()
+    public function testInstanceOfConsoleCommand(): void
     {
-        $this->assertInstanceOf('\Symfony\Component\Console\Command\Command', $this->command);
+        static::assertInstanceOf(Command::class, $this->command);
     }
 
-    public function testBaseCommandClassIsCalled()
+    public function testBaseCommandClassIsCalled(): void
     {
-        $this->assertEquals($this->commandName, $this->command->getName());
+        static::assertSame($this->commandName, $this->command->getName());
     }
 
-    public function testInstanceOfBackgroundCommand()
+    public function testInstanceOfBackgroundCommand(): void
     {
-        $this->assertInstanceOf('\Phlib\ConsoleProcess\Command\BackgroundCommand', $this->command);
+        static::assertInstanceOf(BackgroundCommand::class, $this->command);
     }
 
-    public function testPidFileOptionIsAdded()
+    public function testPidFileOptionIsAdded(): void
     {
-        $this->assertTrue($this->command->getDefinition()->hasOption('pid-file'));
+        static::assertTrue($this->command->getDefinition()->hasOption('pid-file'));
     }
 
-    public function testActionArgumentIsAdded()
+    public function testActionArgumentIsAdded(): void
     {
-        $this->assertTrue($this->command->getDefinition()->hasArgument('action'));
+        static::assertTrue($this->command->getDefinition()->hasArgument('action'));
     }
 
-    public function testDaemonOptionIsAdded()
+    public function testDaemonOptionIsAdded(): void
     {
-        $this->assertTrue($this->command->getDefinition()->hasOption('daemonize'));
+        static::assertTrue($this->command->getDefinition()->hasOption('daemonize'));
     }
 
-    /**
-     * @expectedException \RuntimeException
-     */
-    public function testForkFails()
+    public function testForkFails(): void
     {
+        $this->expectException(\RuntimeException::class);
+
         $this->setupStartFunctions(-1);
-        $pcntl_signal = $this->getFunctionMock('\Phlib\ConsoleProcess\Command', 'pcntl_signal');
-        $pcntl_signal->expects($this->any())
-            ->will($this->returnValue(true));
+        $pcntl_signal = $this->getFunctionMock(__NAMESPACE__, 'pcntl_signal');
+        $pcntl_signal->expects(static::any())
+            ->willReturn(true);
 
         $this->tester->execute([
             'action' => 'start',
-            '-p'     => '/path/to/my.pid',
-            '-d'     => true
+            '-p' => '/path/to/my.pid',
+            '-d' => true,
         ]);
     }
 
-    /**
-     * @expectedException \RuntimeException
-     */
-    public function testChildFailsToGetSession()
+    public function testChildFailsToGetSession(): void
     {
+        $this->expectException(\RuntimeException::class);
+
         $this->setupStartFunctions(null, -1);
-        $pcntl_signal = $this->getFunctionMock('\Phlib\ConsoleProcess\Command', 'pcntl_signal');
-        $pcntl_signal->expects($this->any())
-            ->will($this->returnValue(true));
+        $pcntl_signal = $this->getFunctionMock(__NAMESPACE__, 'pcntl_signal');
+        $pcntl_signal->expects(static::any())
+            ->willReturn(true);
 
         $this->tester->execute([
             'action' => 'start',
-            '-p'     => '/path/to/my.pid',
-            '-d'     => true
+            '-p' => '/path/to/my.pid',
+            '-d' => true,
         ]);
     }
 
-    public function testChildExecutesSuccessfully()
+    public function testChildExecutesSuccessfully(): void
     {
         $expected = 'execute called';
         $this->command->setExecuteOutput($expected);
         $this->setupStartFunctions(null);
-        $pcntl_signal = $this->getFunctionMock('\Phlib\ConsoleProcess\Command', 'pcntl_signal');
-        $pcntl_signal->expects($this->any())
-            ->will($this->returnValue(true));
+        $pcntl_signal = $this->getFunctionMock(__NAMESPACE__, 'pcntl_signal');
+        $pcntl_signal->expects(static::any())
+            ->willReturn(true);
 
         $this->tester->execute([
             'action' => 'start',
-            '-p'     => '/path/to/my.pid',
-            '-d'     => true
+            '-p' => '/path/to/my.pid',
+            '-d' => true,
         ]);
-        $this->assertContains("$expected\n", $this->tester->getDisplay());
+        static::assertStringContainsString("${expected}\n", $this->tester->getDisplay());
     }
 
-    public function testChildCallsOnShutdown()
+    public function testChildCallsOnShutdown(): void
     {
         $expected = 'onShutdown called';
         $this->command->setShutdownOutput($expected);
         $this->setupStartFunctions(null);
-        $pcntl_signal = $this->getFunctionMock('\Phlib\ConsoleProcess\Command', 'pcntl_signal');
-        $pcntl_signal->expects($this->any())
-            ->will($this->returnValue(true));
+        $pcntl_signal = $this->getFunctionMock(__NAMESPACE__, 'pcntl_signal');
+        $pcntl_signal->expects(static::any())
+            ->willReturn(true);
 
         $this->tester->execute([
             'action' => 'start',
-            '-p'     => '/path/to/my.pid',
-            '-d'     => true
+            '-p' => '/path/to/my.pid',
+            '-d' => true,
         ]);
-        $this->assertContains("$expected\n", $this->tester->getDisplay());
+        static::assertStringContainsString("${expected}\n", $this->tester->getDisplay());
     }
 
-    public function testStoppingSuccessfully()
+    public function testStoppingSuccessfully(): void
     {
         $expected = 231;
         $this->setupStopFunctions($expected);
 
-        $posix_kill = $this->getFunctionMock('\Phlib\ConsoleProcess\Command', 'posix_kill');
-        $posix_kill->expects($this->atLeast(2))->with($this->equalTo($expected))->willReturn(false);
+        $posix_kill = $this->getFunctionMock(__NAMESPACE__, 'posix_kill');
+        $posix_kill->expects(static::atLeast(2))
+            ->with($expected)
+            ->willReturn(false);
 
         $this->tester->execute([
             'action' => 'stop',
-            '-p'     => '/path/to/my.pid'
+            '-p' => '/path/to/my.pid',
         ]);
     }
 
-    protected function setupStartFunctions($fork, $setsid = 0, $fexists = false, $writeable = true, $putContents = true)
-    {
-        $namespace = '\Phlib\ConsoleProcess\Command';
+    protected function setupStartFunctions(
+        ?int $fork,
+        int $setsid = 0,
+        bool $fexists = false,
+        bool $writeable = true,
+        bool $putContents = true
+    ): void {
+        $pcntl_fork = $this->getFunctionMock(__NAMESPACE__, 'pcntl_fork');
+        $pcntl_fork->expects(static::any())->willReturn($fork);
 
-        $pcntl_fork = $this->getFunctionMock($namespace, 'pcntl_fork');
-        $pcntl_fork->expects($this->any())->willReturn($fork);
+        $posix_setsid = $this->getFunctionMock(__NAMESPACE__, 'posix_setsid');
+        $posix_setsid->expects(static::any())->willReturn($setsid);
 
-        $posix_setsid = $this->getFunctionMock($namespace, 'posix_setsid');
-        $posix_setsid->expects($this->any())->willReturn($setsid);
+        $file_exists = $this->getFunctionMock(__NAMESPACE__, 'file_exists');
+        $file_exists->expects(static::any())->willReturn($fexists);
 
-        $file_exists = $this->getFunctionMock($namespace, 'file_exists');
-        $file_exists->expects($this->any())->willReturn($fexists);
+        $is_writable = $this->getFunctionMock(__NAMESPACE__, 'is_writable');
+        $is_writable->expects(static::any())->willReturn($writeable);
 
-        $is_writable = $this->getFunctionMock($namespace, 'is_writable');
-        $is_writable->expects($this->any())->willReturn($writeable);
+        $is_writable = $this->getFunctionMock(__NAMESPACE__, 'file_put_contents');
+        $is_writable->expects(static::any())->willReturn($putContents);
 
-        $is_writable = $this->getFunctionMock($namespace, 'file_put_contents');
-        $is_writable->expects($this->any())->willReturn($putContents);
+        $is_writable = $this->getFunctionMock(__NAMESPACE__, 'unlink');
+        $is_writable->expects(static::any())->willReturn(true);
 
-        $is_writable = $this->getFunctionMock($namespace, 'unlink');
-        $is_writable->expects($this->any())->willReturn(true);
-
-        $is_writable = $this->getFunctionMock($namespace, 'sleep');
-        $is_writable->expects($this->any())->willReturn(true);
+        $is_writable = $this->getFunctionMock(__NAMESPACE__, 'sleep');
+        $is_writable->expects(static::any())->willReturn(true);
     }
 
-    protected function setupStopFunctions($pid, $fexists = true, $opened = true, $withPosixKill = false)
-    {
-        $namespace = '\Phlib\ConsoleProcess\Command';
+    protected function setupStopFunctions(
+        int $pid,
+        bool $fexists = true,
+        bool $opened = true,
+        bool $withPosixKill = false
+    ): void {
+        $file_exists = $this->getFunctionMock(__NAMESPACE__, 'file_exists');
+        $file_exists->expects(static::any())->willReturn($fexists);
 
-        $file_exists = $this->getFunctionMock($namespace, 'file_exists');
-        $file_exists->expects($this->any())->willReturn($fexists);
+        $fopen = $this->getFunctionMock(__NAMESPACE__, 'fopen');
+        $fopen->expects(static::any())->willReturn($opened);
 
-        $fopen = $this->getFunctionMock($namespace, 'fopen');
-        $fopen->expects($this->any())->willReturn($opened);
+        $fgets = $this->getFunctionMock(__NAMESPACE__, 'fgets');
+        $fgets->expects(static::any())->willReturn($pid);
 
-        $fgets = $this->getFunctionMock($namespace, 'fgets');
-        $fgets->expects($this->any())->willReturn($pid);
-
-        $fclose = $this->getFunctionMock($namespace, 'fclose');
-        $fclose->expects($this->any())->willReturn(true);
+        $fclose = $this->getFunctionMock(__NAMESPACE__, 'fclose');
+        $fclose->expects(static::any())->willReturn(true);
 
         if ($withPosixKill) {
-            $posix_kill = $this->getFunctionMock($namespace, 'posix_kill');
-            $posix_kill->expects($this->any())->willReturn(false);
+            $posix_kill = $this->getFunctionMock(__NAMESPACE__, 'posix_kill');
+            $posix_kill->expects(static::any())->willReturn(false);
         }
 
-        $usleep = $this->getFunctionMock($namespace, 'usleep');
-        $usleep->expects($this->any())->willReturn(true);
+        $usleep = $this->getFunctionMock(__NAMESPACE__, 'usleep');
+        $usleep->expects(static::any())->willReturn(true);
     }
 }