+++ tests/Command/Stub/ExecuteStubTrait.php
@@ -1,11 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
+namespace Phlib\ConsoleProcess\Command\Stub;
+
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * Class ExecuteStubTrait
- */
 trait ExecuteStubTrait
 {
     /**
@@ -13,22 +14,17 @@ trait ExecuteStubTrait
      */
     protected $executeValue = null;
 
-    /**
-     * @inheritdoc
-     */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        if (!is_null($this->executeValue)) {
+        if ($this->executeValue !== null) {
             $output->writeln($this->executeValue);
         }
         $this->shutdown();
+
+        return 0;
     }
 
-    /**
-     * @param string $value
-     * @return $this
-     */
-    public function setExecuteOutput($value)
+    public function setExecuteOutput(string $value): self
     {
         $this->executeValue = $value;
         return $this;